requirement.txt
# Base requirements
Flask>=2.1.2
Flask-Login>=0.6.1
Flask-SocketIO>=5.1.2
mongoengine>=0.24.1
dnspython>=2.2.1
pycryptodomex>=3.11.0

# For deployment
redis>=4.3.1
gevent>=21.12.0
gunicorn>=20.1.0